backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
python-dotenv
psycopg2-binary
python-multipart